Provide base for fixing issue MetaModels/core#607.
The name of the sub widget is not available in the event and the only way to
bind the options event directly to a sub-widget was by using the event-
propagator.
This event propagator is now being deprecated and will get removed from
dc-general soon and therefore a replacement was needed.
This introduces a new event which will get fired by the MCW to retrieve
options.